Split testing.py into file in directory structure
