+++ test/test_grounding.py
@@ -6,10 +6,9 @@ from feature_extraction import SimpleFeatureExtractor
 from classification import KNNClassifier
 from ocr import OCR, reconstruct_chars
 import mock
-import opencv_utils
 
 
-class Testing(unittest.TestCase):
+class TestGrounding(unittest.TestCase):
     def test_grounding_digits(self):
         grounder = TextGrounder()
         self.assertTrue(grounder)
@@ -75,46 +74,3 @@ class Testing(unittest.TestCase):
                                                      "07816406286208998628034825342117067982148086513282306647093844609"
                                                      "55058223172535940812848111745028410270193852110555964462294895493"
                                                      "038196442881097566593344612847")
-
-
-class TestOCR(unittest.TestCase):
-    def test_ocr_digits(self):
-        segmenter = ContourSegmenter()
-        self.assertTrue(segmenter)
-        extractor = SimpleFeatureExtractor()
-        self.assertTrue(extractor)
-        classifier = KNNClassifier()
-        self.assertTrue(classifier)
-        ocr = OCR(segmenter, extractor, classifier)
-        self.assertTrue(ocr)
-        ocr.train(ImageFile('digits1'))
-        digits = ImageFile('digits2')
-        self.assertTrue(digits)
-        classes, segments = ocr.ocr(digits, show_steps=False)
-        self.assertEqual(reconstruct_chars(classes), "31415926535897932384626433832795028841971693993751058209749445923"
-                                                     "07816406286208998628034825342117067982148086513282306647093844609"
-                                                     "55058223172535940812848111745028410270193852110555964462294895493"
-                                                     "038196442881097566593344612847")
-
-
-class TestImprovements(unittest.TestCase):
-    def test_opencv_brightness_raise(self):
-        image = ImageFile('digits1')
-        processor = opencv_utils.BrightnessProcessor(brightness=2.0)
-        self.assertRaises(AssertionError, lambda: processor._process(image.image))
-
-    def test_opencv_brightness(self):
-        image = ImageFile('digits1')
-        processor = opencv_utils.BrightnessProcessor(brightness=0.5)
-        processor._process(image.image)
-        # TODO: Add checking and try display() function
-        # TODO: Verify the result
-
-    # TODO: Check other ImageProcessors
-
-    def test_opencv_imageprocesser(self):
-        processor = opencv_utils.ImageProcessor()
-        self.assertRaises(NotImplementedError, lambda: processor._image_processing(object))
-
-unittest.main()
-
